Renames Pagination2 to Pagination thus replaces old version with the refactored version
