+++ src/components/Grid/Pagination.js
@@ -1,61 +1,42 @@
 import PropTypes from 'prop-types'
-import React, { memo } from 'react'
-import styles from './Pagination.module.scss'
+import { memo } from 'react'
+
 import { usePagination } from './usePagination'
 
-export const Pagination = memo(({ fetchPageCallback, renderCallback }) => {
-  if (!fetchPageCallback || !renderCallback) {
-    throw Error(
-      'Pagination requires a fetchPageCallback and renderCallback parameters'
-    )
-  }
-
-  const { fetchPage, pagingState, getPaginationNumbers } = usePagination({
-    fetchPage: fetchPageCallback,
-  })
-
-  const conditionallyRenderPagingButtons = () => {
-    if (pagingState.pageCount > 1) {
-      return (
-        <nav aria-label="pagination" className={styles.pagination}>
-          <button
-            className={[
-              styles.paginationButtons,
-              styles.paginationButtonsLeft,
-            ].join(' ')}
-            onClick={() => fetchPage(1)}
-            aria-label="Goto Page 1"
-          >
-            &laquo;
-          </button>
-          {getPaginationNumbers()}
-          <button
-            className={[
-              styles.paginationButtons,
-              styles.paginationButtonsRight,
-            ].join(' ')}
-            onClick={() => fetchPage(pagingState.pageCount)}
-            aria-label={`Goto Page ${pagingState.pageCount}`}
-          >
-            &raquo;
-          </button>
-        </nav>
+/**
+ * Pagination component for use with the data grid
+ *
+ * @public (or @private?)
+ *
+ * @param {object} props - Component Props
+ * @prop {number} [props.currentPage] - The current page
+ * @prop {number} [props.pageCount] - The total number of pages available for the collection of things
+ * @prop {function} [props.fetchPageCallback] - Callback thats fires when the next page should be fetched
+ *
+ * @return {JSX.Element}
+ */
+export const Pagination = memo(
+  ({ currentPage, pageCount, fetchPageCallback }) => {
+    if (
+      typeof currentPage !== 'number' ||
+      typeof pageCount !== 'number' ||
+      !fetchPageCallback
+    ) {
+      throw Error(
+        'Pagination requires currentPage, pageCount, and fetchPageCallback parameters'
       )
-    } else {
-      return null
     }
-  }
+    const { conditionallyRenderPagingButtons } = usePagination({
+      fetchPageCallback,
+    })
 
-  return (
-    <>
-      {renderCallback(pagingState.items)}
-      {conditionallyRenderPagingButtons()}
-    </>
-  )
-})
+    return conditionallyRenderPagingButtons(pageCount, currentPage)
+  }
+)
 
 Pagination.propTypes = {
   fetchPageCallback: PropTypes.func.isRequired,
-  renderCallback: PropTypes.func.isRequired,
+  currentPage: PropTypes.number.isRequired,
+  pageCount: PropTypes.number.isRequired,
 }
 Pagination.displayName = 'Pagination'
