+++ src/components/Grid/index.js
@@ -2,5 +2,4 @@ export { Grid } from './Grid'
 export { Row } from './Row'
 export { Column } from './Column'
 export { Pagination } from './Pagination'
-export { Pagination2 } from './Pagination2'
 export { useGridSorting } from './useGridSorting'
